remove unused parameter in checkRule function
rules are lambdas which compare a card to a stack and return true if the card is allowed onto that stack or false if the card is not allowed onto that stack. Checking the rule however, doesnt need to take the stack as an attribute since the stack stores its rule function, so the stack in question is always the object itself. Therefore, checking a rule should only take a card as a parameter, since the stack to check against is implied when the function is called.